fix: migration expense

+++ database/migrations/2024_12_30_110440_rename_expense_items_table.php
@@ -12,7 +12,10 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('expense_items', function(Blueprint $table) {
-            $table->dropColumn('price');
+            if (Schema::hasColumn('expense_items', 'price')) {
+                $table->dropColumn('price');
+            }
+            $table->tinyInteger('payment_status');
             $table->text('notes')->nullable()->after('total_price');
         });
 
@@ -25,7 +28,8 @@ return new class extends Migration
             $table->bigInteger('price');
             $table->text('notes')->nullable();
 
-            $table->foreign('expense_id')->references('expense_id')->on('expenses')->onUpdate('no action')->onDelete('no action');
+            $table->foreign('expense_id')->references('expense_id')->on('expenses')
+                ->onUpdate('no action')->onDelete('no action');
         });
     }
 
@@ -35,11 +39,19 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('main_prices', function(Blueprint $table) {
-            $table->bigInteger('price');
-            $table->dropColumn('notes');
+            if (Schema::hasColumn('main_prices', 'payment_status')) {
+                $table->dropColumn('payment_status');
+            }
+            if (Schema::hasColumn('main_prices', 'notes')) {
+                $table->dropColumn('notes');
+            }
+            if (Schema::hasColumn('main_prices', 'price')) {
+                $table->dropColumn('price');
+            }
         });
-        Schema::rename('main_prices', 'expense_items');
-        Schema::dropIfExists('expense_addit_prices');
 
+        Schema::rename('main_prices', 'expense_items');
+
+        Schema::dropIfExists('expense_addit_prices');
     }
 };
